Beam-Limiting Devices

Jaws

Calling plot_bld with jaws will plot a simple rectangle where the jaw positions lie,

    plot_bld(jaws)

RoentgenPlots.plot_bld! — Method

plot_bld!(p, jaws::Jaws; kwargs...)

Plot the jaw positions.

kwargs... are passed to Plots.jl plot

Axes Limits

As well as plotting the positions of the jaws, the axis limits can also be set based on the jaw positions,

    axes_lims!(jaws)

RoentgenPlots.axes_lims! — Method

axes_lims!(p, jaws::Jaws; pad=10.)

Set the axes limits to the position of the jaws.

By default, a pad of 10 mm is added to each side.

Multi-Leaf Collimator

Calling plot_bld! with an MLC fills in the region obscured by the leaves.

plot_bld(mlc)

RoentgenPlots.plot_bld! — Method

plot_bld!(p, mlc::MultiLeafCollimator; kwargs...)

Plot the aperture of a MultiLeafCollimator with leaf positions.

Fills the area obstructed by leaves, unless invert=true where it fills the open aperture.

Args:

invert: Fills the region not obscured by the leaves, defaults to false

fill: Whether to fill, defaults to true

fillalpha: Transparency of the fill, defaults to 0.1

leaf_length: The length of the leaves, defaults to 125 mm

Further kwargs... are passed to Plots.jl plot()

Here, plot_bld! is used to add to an existing figure:

plot_bld(mlc; label="Default")
plot_bld!(mlc+(25., 30.); invert=true, fillalpha=0.8, label="Inverted")
plot_bld!(mlc-(50., 50.); fill=false, label="No fill")