Bixels

Bixels are 2D rectangular elements defined on the isoplane. They are also plotted using plot_bld,

plot_bld(Bixel(0., 0., 1., 2.))
plot_bld!(Bixel(10., 5., 3., 2.))

RoentgenPlots.plot_bld! — Method

plot_bld!(p, bixel::AbstractBixel[, value]; kwargs...)

Plot a bixel.

To colour the bixel by a value (e.g. for a beamlet weight), can pass an optional value argument.

The value of the bixel can also be plotted with a colourbar,

plot_bld(Bixel(0., 0., 1., 2.), 0.5)
plot_bld!(Bixel(10., 5., 3., 2.), 1.)
plot_bld!(Bixel(-10., 2., 1., 1.), -0.1)

Collections of Bixels

As well as plotting individual bixels, a collection of bixels can be plotted using the same syntax,

plot_bld(bixels)

RoentgenPlots.plot_bld! — Method

plot_bld!(p, bixel::AbstractArray{<:AbstractBixel}[, value::AbstractArray]; kwargs...)

Plot a collection of bixels.

To colour the bixels by a value (e.g. for a beamlet weight), can pass an optional value array.

A vector of values (w) can also be supplied to colour the bixels,

plot_bld(bixels, w)